pyproject.toml
[tool.poetry]
name = "planning-server"
version = "0.1.0"
description = ""
authors = ["Christopher Hamilton <chrish001@live.com.au>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
strawberry-graphql = {extras = ["debug-server"], version = "^0.220.0"}
fastapi = "^0.110.0"
pymongo = "^4.6.2"
uvicorn = {extras = ["standard"], version = "^0.28.0"}


[tool.poetry.group.dev.dependencies]
mypy = "^1.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
